requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-countries==7.5.1
django-import-export==3.2.0
django-jazzmin==2.6.0
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
lxml==4.9.2
MarkupPy==1.14
numpy==1.24.3
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.0.1
Pillow==9.5.0
pkg_resources==0.0.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.9.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tablib==3.4.0
tinycss2==1.2.1
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.2
webencodings==0.5.1
wget==3.2
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
